Corrected issue if an exception was raised before publishing anync event.  This was resulting the exception not to be raised to the caller due to a null reference exception.
Added the InProcessHandler attribute used to specify in-process event handlers.  This makes it more deterministic and provides a 1-to-1 mapping between the Publisher that dispatches the event.

Corrected issue where the result of a command was not being automatically set.  Now the event handler don't needs to set the response on the event being handled - It just returns the result from the handler.

Corrected Namespaces for the NetFusion.Settings.MongoDb assembly that were named incorrectly.

Renamed each plug-in manifest derived class to contain the name of the plug-in within its name.