Configuration Updated and removal of unused projects.
Removed one project that was empty to be used for future implementations and deleted a logging project that is no longer needed.  Make small change to configuration so the host application now specifics an IConfiguration to the contianer during bootstrap and not IConfigurationBuilder.  This allows the host to access the configuration (i.e. when configuring logging).