Refactor: Moved Query implementation to NetFusion.Messaging
The Query implementation has been moved to NetFusion.Messaging.  It use to be loaded in the NetFusion.Domain.Patterns.  However, the Query implementation is no more a pattern than the Command and DomainEvent.  This also prevents the consuming application to have a reference on NetFusion.Domain.Patterns since this should be considered an optional assembly.  The NetFusion.Domain.Patterns assembly will now only contain more specific patterns such as the UnitOfWork....  Also, the method to dispatch a Query has been consolidated into the IMessagingService.