Composite Log Updates No Breaking Changes...
Just better naming the items that are added to the Composite log.   Also extended, the log for the concrete types contained within a plug-in that are based on abstract types contained in another.  This gives a complete view of how the application was constructed.  Not breaking changes since this is just a nested dictionary of key/value pairs to record how a project was build.