Minor refactor to logging + Bug fix...
Modified the LogDetail MS ILogger extensions that add to the base set of log methods provided by MS.  These are the methods that accept a detail object that is serialized to JSON and included as part o fthe log.  Modified all of the methods declarations to only take the detail object and removed the arguments parameter.  Don't see a need for both and this address the bug that was discovered in the logging.  Also changing build file so a commit will not be triggered for a change to read-me files.