Improved logging of invalid MS loaded configurations.
Change code so that it first checks if the application/configuration setting is invalid after having the configuration loaded from MS configuration settings.  If invalid, an error is logged with the setting type, configuration section name, and the JSON serialization of the invalid state.  Then an exception is thrown as was already being done.