Created RequestClientBuilder
This new class will replace the current RequestClientFactory.  The builder provides improved support for loading the initial service api exposed by a service.  It also included a delegate that can be specified that will be invoked on each request.  Logging will be added next after class is testing in Boondocks.