First set of refactors to provide easy integration with .net core 3.0 and the current GenericHost. The next version of .net core has a breaking change in terms of how applications are initialized and built.  No longer will you be able to inject ILogger and IConfiguration into the Startup classes.  These services are only available in the Configure method of the startup class.  MS made these changes to support only one form of application initialization (Generic Host). The prior version of .net core was creating a temporary container that was being used to DI the Startup class.  Now there will be only one created container.
The CompositeContainer is now just a means to the end.  It is used to bootstrap the application but never used or available afterwards.  The CompositeContainer delegates to the CompositeAppBuiler that now does most of the code that was in the CompositeContainer.  The CompositeAppBuilder at the end of its execution registers a new class called CompositeApp within the container as a singleton.  This represents the complete application that was composed by registered plugins.