Broke out the classes used by unit-tests...
This was required since having ASP.NET classes such as controllers within a test project was resulting in a warning dealing with Analyzers.  The error was a null reference exception for "top level controller parameters).  This is a known issue if multiple .net runtimes are referenced but this is not the case here.  However, moving the controller classes out of the unit-test project and into a class assembly fixes the issue.