Updated .NET Core Microservice Template CLI
Updated the generated Dockerfile so it now uses the correct base images for .NET Core 3.1 SDK and runtime version of the new aspnet core image.  Moved the Solution file down a directory so it is within the child src directory.  Also removed the scripts and VSCode task for debugging a container running in Docker.  Really have yet to have need for this so removing since it only works on linux.