Simplifying base message attributes...
The base message Attributes property is now IDictionary<string, string> instead of IDictionary<string, object>.  The reason this has been changed is so the implementation is independent of the type of serializer being used.  On the receiving end, when the message properties are deserialized, the type of the originally serialized value is not know since the dictionary is of type object.  The NewtonSoft serializer will "guess" the type but it might not be exactly the type was when serialized.  For example, the NewtonSoft serializer detects that it is a number and deserializes to a Long even if an Int was used when serialized.  Other serializers, such as the new MS Json serializer and MessagePack do not attempt to determine the type.  Instead, both deserialize the value into a wrapper object defined by its code-base.  While the value can be extracted, this would have to be done for each type of serializer.  Since these are just key/values pairs that are only used to communicate message properties, the code has been refactored to use IDictionary<string, string> which will always deserialze to a string.  Then a set of extension methods for the most common types have been provided to set/get values parsed from the string.  This is a lower level concern used mostly by the implementation so have a consistent way to store the values that works across all serializers is best.