Additional logging refactoring
Added the start of the log extensions that delegate to Serilog.  Removed all log event Id values.  Have not found this very useful.  Next, will go through the bootstrap logic and create grouped log messages so log will only show message with a summary.  Then all the details are logged as associated properties.  This will greatly reduce the number of log messages and make it easier to view.