Added ITypeLog
This interface is checked when logging the state of an object.  The type being logged can implement this interface to reduce the amount of data that is logged.  If not implemented, the entire state of the object is logged.  If the object being logged is small, this interface need not be implemented.