Updated all projects to .net 6
There was an odd bug introduced for which the code was modfied within the SettingsModule to correct.  Since IOption<T> is generic and does not have a base non-generic interface, the object of this type was being access via dynamic reference to retrieve the Value property.  For some reason this is not working after updating the nugets so swtiched code to use reflection.